temporary

+++ psx-modding-toolchain/games/CrashTeamRacing/decompile/General/DecalHUD/DecalHUD_DrawPolyGT4 - Readable.c
@@ -1,6 +1,6 @@
 #include <common.h>
 
-void DECOMP_DecalHUD_DrawPolyGT4(struct Icon* icon, int posX, int posY, struct PrimMem* primMem, u_long* ot, u_int color0, u_int color1, u_int color2, u_int color3, char transparency, short scale)
+void DECOMP_DecalHUD_DrawPolyGT4(struct Icon* icon, short posX, short posY, struct PrimMem* primMem, u_long* ot, u_int color0, u_int color1, u_int color2, u_int color3, char transparency, int scale)
 {
 	short width;
 	short height;
